Stop accepting 503 responses to an EHLO attempt.
Servers that only speak HELO might return a 503 status code to
indicate that EHLO is unsupported.  In this case, we want the HELO
fallback code to run instead.

I'm not exactly sure what case the previous code was attempting to
address, but it looks like we already catch 503 responses in the
individual authentication methods, so perhaps this is no longer
necessary.

If this change results in a regression, we'll need to address the root
issue another way.

Closes #5